Moving functions to helper function script
